+++ core/server/api_container/server/startosis_engine/startosis_packages/git_package_content_provider/locators.go
@@ -18,8 +18,12 @@ func isLocalLocator(locator string) bool {
 	return false
 }
 
-func isSamePackageLocalAbsoluteLocator(locator string, parentPackageId string) bool {
-	return strings.HasPrefix(locator, parentPackageId)
+func shouldBlockAbsoluteLocatorBecauseIsInTheSameSourceModuleLocatorPackage(relativeOrAbsoluteLocator string, sourceModuleLocator string, rootPackageId string) bool {
+
+	isSourceModuleInRootPackage := strings.HasPrefix(sourceModuleLocator, rootPackageId)
+	isAbsoluteLocatorInRootPackage := strings.HasPrefix(relativeOrAbsoluteLocator, rootPackageId)
+
+	return isSourceModuleInRootPackage && isAbsoluteLocatorInRootPackage
 }
 
 func replaceAbsoluteLocator(absoluteLocator string, packageReplaceOptions map[string]string) string {