+++ internal_testsuites/golang/testsuite/startosis_idempotency_test/no_instruction_cached_test.go
@@ -1,4 +1,4 @@
-package startotis_idempotent_run_test
+package startosis_persistent_directory_test
 
 import (
 	"context"